orbits fetched and cached from basevis (currently the callback is on completed tick)
